capture: use uint16_t for max_pending_packets
Use a fixed type of max_pending_packets instead of intmax_t which can
differ based on the platform/standard library.

Should also prevent lints about possible arithmetic overflow.

+++ src/source-pcap-file.c
@@ -32,7 +32,7 @@
 #include "runmode-unix-socket.h"
 #include "suricata.h"
 
-extern intmax_t max_pending_packets;
+extern uint16_t max_pending_packets;
 PcapFileGlobalVars pcap_g;
 
 /**
